Created library folder
Moved all library related components into library folder. Refactored library page into smaller components
