+++ src/components/library/LibraryActions.js
@@ -5,21 +5,26 @@ import { MdInfoOutline } from "react-icons/md";
 import Modal from "../../helpers/modal/Modal";
 import AddToLibrary from "./AddToLibrary";
 import { AuthContext } from "../../contexts/authContext";
+import Login from "../login/Login";
 
 const LibraryActions = (props) => {
   const navigate = useNavigate();
   const { currentUser } = useContext(AuthContext);
   const [openModal, setOpenModal] = useState(false);
+  const [openLoginModal, setOpenLoginModal] = useState(false);
 
   //navigate to the book details page for the specified book
   const handleDetails = () => navigate(`/results/${props.book.id}`);
 
-  //set the modal state when the user clicks on the add to library button
+  /**
+   * if user is signed in, open the add to library modal
+   * if user is not signed in, open the login modal
+   */
   const handleLibrary = () => {
     if (currentUser.email) {
       setOpenModal((state) => !state);
     } else {
-      alert("You must be logged in to add a book to your library");
+      setOpenLoginModal((state) => !state);
     }
   };
 
@@ -49,6 +54,12 @@ const LibraryActions = (props) => {
       {openModal && (
         <Modal setOpenModal={setOpenModal} openModal={openModal}>
           <AddToLibrary selectedBook={props.book} setOpenModal={setOpenModal} />
+        </Modal>
+      )}
+
+      {openLoginModal && (
+        <Modal setOpenModal={setOpenLoginModal} openModal={openLoginModal}>
+          <Login setOpenModal={setOpenLoginModal} />
         </Modal>
       )}
     </>
