+++ src/components/pagesStyles/Library.module.css
@@ -1,60 +1,3 @@
 .library-container {
   padding-top: 80px;
 }
-
-.library-navbar {
-  margin-top: 5rem;
-  padding: 1.5rem 0;
-  border-bottom: 1.5px solid rgba(166, 166, 166, 0.3);
-  margin-bottom: 2rem;
-}
-
-.library-navbar h1 {
-  display: none;
-}
-
-.library-links {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  /* padding-top: 1rem; */
-  cursor: pointer;
-}
-
-.link {
-  color: var(--med-grey);
-  font-weight: 400;
-}
-
-.active {
-  font-weight: 600;
-  color: var(--dark-blue);
-}
-
-/* Media Queries */
-
-@media (min-width: 768px) {
-  .library-navbar h1 {
-    display: grid;
-    color: var(--dark-blue);
-    font-weight: bold;
-  }
-
-  .library-navbar {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding-top: 1.5rem;
-  }
-
-  .library-links {
-    padding-top: 0;
-    gap: 1.5rem;
-  }
-}
-
-@media (min-width: 900px) {
-  .library-navbar {
-    margin-top: 0;
-  }
-}
