+++ src/components/search/SearchResults.js
@@ -4,7 +4,7 @@ import { Navigate, useLocation } from "react-router-dom";
 import Books from "../books/Books";
 import Container from "../../helpers/wrapper/Container";
 import Heading from "../../helpers/heading/Heading";
-import LibraryActions from "../books/LibraryActions";
+import LibraryActions from "../library/LibraryActions";
 import Loading from "../../helpers/Loading";
 import EmptyShelf from "../books/EmptyShelf";
 import webSearch from "../../images/web_search.svg";
